fix synth harness generation for axi

+++ src/unit_test_helpers/create_synthesis_harness.py
@@ -38,8 +38,13 @@ def get_testbench_dependencies(RTL_DIR, flist):
             all_files = get_all_files_in_dir(dir)
             dependencies_list.extend(all_files)
 
-        dependencies_list += get_all_files_in_dir(RTL_DIR)
+        dependencies_list += [f for f in get_all_files_in_dir(RTL_DIR) if '_lite_' in f ]
         cleaned_dependencies = [dep.strip() for dep in dependencies_list if dep.strip() and not dep.startswith("+incdir+")]
+        cleaned_dependencies = list(set(cleaned_dependencies))
+        for f in cleaned_dependencies:
+            if 'tb' in os.path.basename(f) or 'test' in os.path.basename(f):
+                print(f"Removing testbench file: {f}")
+                cleaned_dependencies.remove(f)
         return cleaned_dependencies
         
 
@@ -53,36 +58,49 @@ def preprocess_files(file):
         content = content.replace('`include', '// `include')
         return content
 
-def create_synthesis_harness(dependencies_list):
-    SYNTH_DIR = os.path.join(CUR_DIR, "synthesis")
+def create_synthesis_harness(dependencies_list,testbench_name):
+    dependencies_list.sort(key=lambda x: (not ((x.endswith('.svh')) or ('pkg' in os.path.basename(x))), x))
+    # print(f"sorted dependencies list: {dependencies_list}")
+    SRC_DIR = os.path.join(CUR_DIR, "../build")
+    os.makedirs(SRC_DIR, exist_ok=True)
+    SYNTH_DIR = os.path.join(SRC_DIR, testbench_name.split('.')[0])
+    if os.path.exists(SYNTH_DIR):
+        print(f"Warning: Directory {SYNTH_DIR} already exists. It will be overwritten.")
+        os.system(f"rm -rf {SYNTH_DIR}")
     os.makedirs(SYNTH_DIR, exist_ok=True)
     os.makedirs(os.path.join(SYNTH_DIR, "rtl"), exist_ok=True)
     print(f"Synthesis directory created at: {SYNTH_DIR}")
     for dep in dependencies_list:
         content = preprocess_files(dep)
-        with open(f"{SYNTH_DIR}/rtl/{os.path.basename(dep)}", 'w') as f:
+        with open(os.path.join(SYNTH_DIR, "rtl", testbench_name), 'a') as f:
+            f.write(f"//=========================File: {os.path.basename(dep)}================\n")
             f.write(content)
-    os.system(f"zip -r synthesis_harness.zip synthesis")
-    os.system(f"rm -rf {SYNTH_DIR}")
+    #     with open(f"{SYNTH_DIR}/rtl/{os.path.basename(dep)}", 'w') as f:
+    #         f.write(content)
+    # os.system(f"zip -r synthesis_harness.zip synthesis")
+    # os.system(f"rm -rf {SYNTH_DIR}")
 
 def main():
-    if len(sys.argv) < 3:
-        print("Usage: python create_synthesis_harness.py <testbench_name> <file_list>")
+    if len(sys.argv) < 4:
+        print("Usage: python create_synthesis_harness.py <testbench_path> <file_list> testbench_name")
         sys.exit(1)
 
     testbench_path = sys.argv[1]
     rtl_path = os.path.abspath(testbench_path)
     flist = sys.argv[2]
+    testbench_name = sys.argv[3]+ ".sv"
 
     dependencies_list = get_testbench_dependencies(rtl_path, flist)
+
+    print(f"Found {len(dependencies_list)} dependencies for the testbench.")
 
 
     if not dependencies_list:
         print("No dependencies found for the testbench.")
         sys.exit(1)
 
-        
-    create_synthesis_harness(dependencies_list)
+
+    create_synthesis_harness(dependencies_list, testbench_name)
     print("Synthesis harness created successfully.")
